+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 dev
 ---
 
+* Added access to warnings in decompilation phases and their emission in the
+  ``decompiler`` tool.
 * Added support for selecting the compiler to be used when compiling C source
   files.
 * Added support for selecting the optimizations to be used when compiling C